LPS-31014 Use FULL-CONTENT instead of ABSTRACT as default display style value for feed URL generation.
The RSS taglib does not include in the feed URL those parameters which have the default value, so that the URL is shorter. For the display style, ABSTRACT was being considered as default RSS feed display style value, but in fact FULL-CONTENT is the default value for all the portlets, so it has been changed to solve the issue. The Asset Publisher has ABSTRACT as default value, but it retrieves the value from the portlet preferences so this change has no effect on this portlet.