LPS-34896 Avoid showing the Global scope twice in the menu
This can only happen as of now for company admins since in that case it's returning an ArrayList and not a UniqueList, but I think it's better to have a more generic if clause. It is important to keep the logic to add the company group because for non company admins which certain permissions the getManageableGroups method does not account for all cases.