LPS-35675 Refactoring and source formatting
I refactored the existing code to have isValid methods for all three types of properties (boolean, range and multiple valid values). I also removed the logging of the case where the rule does not specify the value of a property because I think it's not as useful and adds noise. Finally I created three log methods to simplify and provide more specific methods.