LPS-39662 Read cookies that start with a given prefix and make them available anywhere that the User object is available
This is necessary for example to be able to segment anonymous users. Plugins will be responsible for setting the cookie, but the portal must be able to read this type of cookies (which I'm calling "Remote Preferences") and make them available as part of the User object so that they are easily accessible from from the code.