LPS-40872 (DO NOT MERGE) use ThreadDeath
Shouldn't we be using ThreadDeath to terminate the thread execution once we find an error such the request being recycled?

I have tried this approach but it does not work as expected. At some
point ThreadDeath is treated by tomcat libraries and, I can't tell why
yet, ends up beign catched in catch (Exception) blocks and logged. This
should not be happening.