LPS-46267 Split Settings into Settings and ModifiableSettings.
This will allow us to differentiate between which keys are persisted and which ones are originated from default or virtual values. This functionality is needed for Archived Setups, for detecting overridden default values in configuration actions and probably for the import/export.

The commit is big, but it contains the minimum changes to split Settings into two interfaces and make the portal compile and the tests pass.