LPS-46267 Split Settings into ModifiableSettings (read-write) and Settings (read-only).
This is done to make clear the difference between Settings which only decorate or alter the search strategy of configuration keys (Settings) and those which persist or store the real configuration keys (ModifiableSettings).

Also create a BaseModifiableSettings class and move some methods from BaseSettings into it.

Add new methods to BaseSettings to handle having a parent settings associated to the current settings instance from which to pick default values. We decided to go with the terminology of "parent" for its similitude with how classloaders work.