LPS-44467 Add the concept of userIdAndInstanceId
This concept encloses the full instance id of a portlet in a layout. It could represent the same concept as instanceId, but including the userId. This it necessary because a portlet can be deployed into a customizable region of a layout, so each user will have her own portlet instance whether the portlet is instanceable or not.

For instance, a noninstanceable portlet will have as userIdAndInstanceId:
- if it's deployed on a noncustomizable region of a layout: NULL -> (userId: 0, instanceId: null) -> portletInstanceKey: 33
- if it's deployed on a customizable region of a layout: 1234_ -> (userId: 1234, instanceId: null) -> portletInstanceKey: 33_USER_1234

On the other hand, an instanceable portlet will have as userIdAndInstanceId:
- if it's deployed on a customizable region of a layout: 1234_abcdef32 -> (userId: 1234, instanceId: abcdef32) -> portletInstanceKey: 33_USER_1234_INSTANCE_abcdef32
- if it's deployed on a noncustomizable region of a layout: abcdef32 -> (userId: 0, instanceId: abcdef32) -> portletInstanceKey: 33_INSTANCE_abcdef32