LPS-112372 Introduce InfoLocalizedValue to represent multi-locale values
Two implementations are provided one based on a builder to which all values are provided and another one which can look for translations in a resource bundle.
Having this class helps solve the long time problem of having to pass the locale to methods just because at some point down the line it's necessary to obtain a value that has different value per locale. Thanks to this class it's possible to return all values and decide later on which locale to pass down if the value is needed at all.